# docs/INF/openapi/teralya-openapi-v1.0.yaml
openapi: 3.1.0
...
info:
  title: Teralya API
  version: 1.0.0
  description: Contrato OpenAPI 3.1 del MVP. Derivado de INF-08 v2.5, INF-10 v1.0 e INF-10-A v1.0.
  contact:
    name: Teralya
    url: https://teralya.es
  license:
    name: Propietario — Teralya
    identifier: LicenseRef-Teralya-Proprietary
servers:
- url: https://api.teralya.es
  description: Producción
...
paths:
...
  /pedidos/{id}:
    get:
      operationId: api020
      x-teralya-api-code: API-020
      x-teralya-cu:
      - CU-011
      x-teralya-pt:
      - PT-COM-007
      tags:
      - Pedidos
      summary: Consultar Pedido propio
      description: Contrato normativo de API-020.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderBuyerDetail'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    numero_pedido: aaaaa
                    estado: pendiente_pago
                    total: '10.00'
                    moneda: EUR
                    created_at: '2026-07-13T12:00:00Z'
                    totales:
                      subtotal: '10.00'
                      gastos_envio: '2.00'
                      impuestos: '2.10'
                      descuentos: '1.00'
                      total: '13.10'
                      moneda: EUR
                    direccion_envio_snapshot:
                      nombre_destinatario: aaaaa
                      direccion: aaaaa
                      codigo_postal: aaaaa
                      ciudad: aaaaa
                      pais: aaaaa
                    direccion_facturacion_snapshot:
                      nombre_destinatario: aaaaa
                      direccion: aaaaa
                      codigo_postal: aaaaa
                      ciudad: aaaaa
                      pais: aaaaa
                    lineas: []
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
        description: Identificador id
        example: 11111111-1111-4111-8111-111111111111
  /bodegas/yo/subpedidos:
    get:
      operationId: api021
      x-teralya-api-code: API-021
      x-teralya-cu:
      - CU-018
      x-teralya-pt:
      - PT-BOD-007
      tags:
      - SubPedidos
      summary: Listar SubPedidos propios
      description: Contrato normativo de API-021.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSubOrderSummary'
              examples:
                success:
                  value:
                    items: []
                    page: 1
                    page_size: 1
                    total_items: 0
                    total_pages: 0
        '400':
          description: Validation Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/validation-error
                    title: Validation Error
                    status: 400
                    detail: La solicitud no pudo completarse.
                    code: VALIDATION_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - *id004
      - *id005
  /bodegas/yo/subpedidos/{id}:
    get:
      operationId: api022
      x-teralya-api-code: API-022
      x-teralya-cu:
      - CU-018
      x-teralya-pt:
      - PT-BOD-008
      tags:
      - SubPedidos
      summary: Consultar SubPedido propio
      description: Contrato normativo de API-022.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubOrderDetail'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    pedido_id: 11111111-1111-4111-8111-111111111111
                    estado: pendiente
                    total: '10.00'
                    moneda: EUR
                    fecha_ultimo_cambio_estado: '2026-07-13T12:00:00Z'
                    totales:
                      subtotal: '10.00'
                      gastos_envio: '2.00'
                      impuestos: '2.10'
                      total: '14.10'
                      moneda: EUR
                    lineas: []
                    direccion_envio_snapshot:
                      nombre_destinatario: aaaaa
                      direccion: aaaaa
                      codigo_postal: aaaaa
                      ciudad: aaaaa
                      pais: aaaaa
                    tracking: {}
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
        description: Identificador id
        example: 11111111-1111-4111-8111-111111111111
  /bodegas/yo/subpedidos/{id}/estado:
    patch:
      operationId: api023
      x-teralya-api-code: API-023
      x-teralya-cu:
      - CU-019
      - CU-030
      x-teralya-pt:
      - PT-BOD-008
      - PT-COM-007
      - PT-ADM-007
      tags:
      - SubPedidos
      summary: Cambiar estado de SubPedido
      description: 'Matriz permitida: pendiente→aceptado|cancelado|incidencia; aceptado→en_preparacion|cancelado|incidencia;
        en_preparacion→enviado|cancelado|incidencia; incidencia→en_preparacion|cancelado; enviado→entregado. Entregado y cancelado
        son terminales; mismo estado es idempotente. El cambio y el recálculo de Pedido son atómicos bajo bloqueo. Derivación:
        todos los no cancelados pendientes→pagado; ninguno enviado/entregado y alguno aceptado/en_preparacion/incidencia→en_preparacion;
        mezcla de enviado/entregado con estados previos→parcialmente_enviado; todos los no cancelados enviado/entregado y
        alguno enviado→enviado; todos los no cancelados entregado→entregado; todos cancelados→cancelado.'
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubOrderDetail'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    pedido_id: 11111111-1111-4111-8111-111111111111
                    estado: pendiente
                    total: '10.00'
                    moneda: EUR
                    fecha_ultimo_cambio_estado: '2026-07-13T12:00:00Z'
                    totales:
                      subtotal: '10.00'
                      gastos_envio: '2.00'
                      impuestos: '2.10'
                      total: '14.10'
                      moneda: EUR
                    lineas: []
                    direccion_envio_snapshot:
                      nombre_destinatario: aaaaa
                      direccion: aaaaa
                      codigo_postal: aaaaa
                      ciudad: aaaaa
                      pais: aaaaa
                    tracking: {}
        '400':
          description: Validation Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/validation-error
                    title: Validation Error
                    status: 400
                    detail: La solicitud no pudo completarse.
                    code: VALIDATION_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '409':
          description: Conflict
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/conflict
                    title: Conflict
                    status: 409
                    detail: La solicitud no pudo completarse.
                    code: CONFLICT
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
        description: Identificador id
        example: 11111111-1111-4111-8111-111111111111
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubOrderStatePatch'
            examples:
              success:
                value:
                  estado_destino: pendiente
...
  /admin/pedidos/{id}:
    get:
      operationId: api039
      x-teralya-api-code: API-039
      x-teralya-cu:
      - CU-025
      x-teralya-pt:
      - PT-ADM-007
      tags:
      - Administración
      summary: Consultar Pedido administrativo
      description: Contrato normativo de API-039.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderAdminDetail'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    numero_pedido: aaaaa
                    estado: pendiente_pago
                    total: '10.00'
                    moneda: EUR
                    created_at: '2026-07-13T12:00:00Z'
                    totales:
                      subtotal: '10.00'
                      gastos_envio: '2.00'
                      impuestos: '2.10'
                      descuentos: '1.00'
                      total: '13.10'
                      moneda: EUR
                    direccion_envio_snapshot:
                      nombre_destinatario: aaaaa
                      direccion: aaaaa
                      codigo_postal: aaaaa
                      ciudad: aaaaa
                      pais: aaaaa
                    direccion_facturacion_snapshot:
                      nombre_destinatario: aaaaa
                      direccion: aaaaa
                      codigo_postal: aaaaa
                      ciudad: aaaaa
                      pais: aaaaa
                    lineas: []
                    subpedidos: []
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
        description: Identificador id
        example: 11111111-1111-4111-8111-111111111111
...
components:
  securitySchemes:
    BearerOpaque:
      type: http
      scheme: bearer
      description: Token opaco de 256 bits; TTL absoluto 8 horas.
  headers:
    XRequestId:
      description: Identificador UUID de trazabilidad
      schema: *id001
  schemas:
    FieldError:
      type: object
      additionalProperties: false
      properties:
        field: &id013
          type: string
          minLength: 1
          maxLength: 200
        code: *id006
        message: &id007
          type: string
          minLength: 1
          maxLength: 500
      required:
      - field
      - code
      - message
    Problem:
      type: object
      additionalProperties: false
      properties:
        type: &id016
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
        title: *id002
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail: *id007
        instance:
          type: string
          minLength: 1
          maxLength: 2048
        code: *id006
        request_id: *id001
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          maxItems: 50
        retryable:
          type: boolean
      required:
      - type
      - title
      - status
      - detail
      - code
      - request_id
      - retryable
...
    MoneyBreakdown:
      type: object
      additionalProperties: false
      properties:
        subtotal: *id003
        gastos_envio: *id003
        impuestos: *id003
        descuentos: *id003
        total: *id003
        moneda: &id008
          type: string
          enum:
          - EUR
      required:
      - subtotal
      - gastos_envio
      - impuestos
      - descuentos
      - total
      - moneda
...
    AddressSnapshot:
      type: object
      additionalProperties: false
      properties:
        nombre_destinatario: *id002
        direccion: *id013
        codigo_postal: *id027
        ciudad: *id009
        pais: *id009
        empresa: *id002
        direccion_adicional: *id013
        provincia: *id009
        persona_contacto: *id009
        telefono: *id014
        email: *id010
      required:
      - nombre_destinatario
      - direccion
      - codigo_postal
      - ciudad
      - pais
...
    OrderLine:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        vino_id: *id001
        nombre_vino: *id002
        bodega: *id002
        precio_unitario: *id003
        cantidad: *id024
        importe_total: *id003
        anada: *id017
      required:
      - id
      - vino_id
      - nombre_vino
      - bodega
      - precio_unitario
      - cantidad
      - importe_total
...
    OrderBuyerDetail:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        numero_pedido:
          type: string
          minLength: 1
          maxLength: 100
        estado:
          type: string
          enum:
          - pendiente_pago
          - pagado
          - en_preparacion
          - parcialmente_enviado
          - enviado
          - entregado
          - cancelado
          - devuelto
        total:
          type: string
          pattern: ^(0|[1-9][0-9]{0,7})\.[0-9]{2}$
        moneda:
          type: string
          enum:
          - EUR
        created_at:
          type: string
          format: date-time
        totales:
          $ref: '#/components/schemas/MoneyBreakdown'
        direccion_envio_snapshot:
          $ref: '#/components/schemas/AddressSnapshot'
        direccion_facturacion_snapshot:
          $ref: '#/components/schemas/AddressSnapshot'
        lineas:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
      required:
      - id
      - numero_pedido
      - estado
      - total
      - moneda
      - created_at
      - totales
      - direccion_envio_snapshot
      - direccion_facturacion_snapshot
      - lineas
    SubOrderStatePatch:
      type: object
      additionalProperties: false
      properties:
        estado_destino: &id030
          type: string
          enum:
          - pendiente
          - aceptado
          - en_preparacion
          - enviado
          - entregado
          - cancelado
          - incidencia
      required:
      - estado_destino
...
    SubOrderSummary:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        pedido_id: *id001
        estado: *id030
        total: *id003
        moneda: *id008
        fecha_ultimo_cambio_estado: *id018
      required:
      - id
      - pedido_id
      - estado
      - total
      - moneda
      - fecha_ultimo_cambio_estado
    SubOrderDetail:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        pedido_id:
          type: string
          format: uuid
        estado:
          type: string
          enum:
          - pendiente
          - aceptado
          - en_preparacion
          - enviado
          - entregado
          - cancelado
          - incidencia
        total:
          type: string
          pattern: ^(0|[1-9][0-9]{0,7})\.[0-9]{2}$
        moneda:
          type: string
          enum:
          - EUR
        fecha_ultimo_cambio_estado:
          type: string
          format: date-time
        totales:
          $ref: '#/components/schemas/SubOrderMoneyBreakdown'
        lineas:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        direccion_envio_snapshot:
          $ref: '#/components/schemas/AddressSnapshot'
        tracking:
          $ref: '#/components/schemas/Tracking'
        pedido_estado: *id029
      required:
      - id
      - pedido_id
      - estado
      - total
      - moneda
      - fecha_ultimo_cambio_estado
      - totales
      - lineas
      - direccion_envio_snapshot
      - tracking
    OrderAdminDetail:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        numero_pedido:
          type: string
          minLength: 1
          maxLength: 100
        estado:
          type: string
          enum:
          - pendiente_pago
          - pagado
          - en_preparacion
          - parcialmente_enviado
          - enviado
          - entregado
          - cancelado
          - devuelto
        total:
          type: string
          pattern: ^(0|[1-9][0-9]{0,7})\.[0-9]{2}$
        moneda:
          type: string
          enum:
          - EUR
        created_at:
          type: string
          format: date-time
        totales:
          $ref: '#/components/schemas/MoneyBreakdown'
        direccion_envio_snapshot:
          $ref: '#/components/schemas/AddressSnapshot'
        direccion_facturacion_snapshot:
          $ref: '#/components/schemas/AddressSnapshot'
        lineas:
          type: array
          items:
            $ref: '#/components/schemas/OrderLine'
        subpedidos:
          type: array
          items:
            $ref: '#/components/schemas/SubOrderDetail'
        comprador_id: *id001
      required:
      - id
      - numero_pedido
      - estado
      - total
      - moneda
      - created_at
      - totales
      - direccion_envio_snapshot
      - direccion_facturacion_snapshot
      - lineas
      - subpedidos
...
    PageSubOrderSummary:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubOrderSummary'
          maxItems: 100
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
      required:
      - items
      - page
      - page_size
      - total_items
      - total_pages
...